pyproject.toml
[tool.poetry]
name = "remind_me"
version = "0.1.0"
description = ""
authors = ["Jesse Brink <jbrink0925@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
APScheduler = "^3.7.0"
python-dateutil = "^2.8.1"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
fastapi-chameleon = {git = "https://github.com/mikeckennedy/fastapi-chameleon", rev = "main"}
python-multipart = "^0.0.5"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
